test(reporters): improve output for test.watch npm script
This makes the output of test.watch more useful by focusing on failing tests. It has some impact on
regular testing in that I could not find a combination of CLI options that would remove test
coverage output. So, I removed the text reporters for code coverage. We still generate artifacts for
it though, so we could update  Jenkins to track and report on it if we like.