fix(stencil-wrapper): fix the stencil-wrapper build step to output to the old location
The previous change to get stencil-wrapper.ts to compile with the import path changes had the
unexpected side-effect of changing where the JS file was output, breaking downstream projects. This
removes the srcDir argument when building, which unforcunately causes TS errors, but doesn't
prevent the correct JS from being generated. I don't like it very much, but it should work as a
stopgap for now.

COMUI-1270